✨ feat(web): embed agent management in the profile page
Extract the settings agent editor into a shared AgentDetailPanel (AgentsPage
1168→109 lines) and embed it in the agent profile page behind the existing
admin-or-creator rule, now centralized as canEditAgent(). Unauthorized users
render no configuration section and issue no prompt-bearing requests (the
settings query is permission-gated via enabled). Adds a read-only About
section for everyone; the skills tab stays a deep link because its two-pane
editor does not fit the profile column.
