+++ web/src/features/agents/AgentForm.tsx
@@ -1,6 +1,5 @@
-import { useNavigate } from "@tanstack/react-router";
 import type { Skill, User } from "@/lib/types";
-import type { AgentsPageState } from "./AgentsPage";
+import type { AgentsPageState } from "./agent-detail-state";
 import { Button } from "@/components/ui/button";
 import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs";
 import { useI18n } from "@/lib/i18n";
@@ -10,12 +9,21 @@ import { SkillsTab } from "./tabs/SkillsTab";
 import { ToolsTab } from "./tabs/ToolsTab";
 import { AdvancedTab } from "./tabs/AdvancedTab";
 import { UsersTab } from "./tabs/UsersTab";
+
+/**
+ * `page` fills a dedicated settings pane and scrolls internally; `embedded`
+ * sits as one card inside a host page that owns the scroll.
+ */
+export type AgentFormLayout = "page" | "embedded";
+
 interface Props {
   state: AgentsPageState;
+  layout?: AgentFormLayout;
+  hiddenTabs?: readonly string[];
   onSetState: (patch: Partial<AgentsPageState>) => void;
+  onTabChange: (tab: string) => void;
   onSave: () => void;
-  onCancel: () => void;
-  onLoadAssignedUsers: (agentId: string) => void;
+  onCancel?: () => void;
   onAddUser: () => void;
   onRemoveUser: (userId: string) => void;
   onApplySoul: (soulID: string) => void;
@@ -30,10 +38,12 @@ interface Props {
 
 export function AgentForm({
   state,
+  layout = "page",
+  hiddenTabs,
   onSetState,
+  onTabChange,
   onSave,
   onCancel,
-  onLoadAssignedUsers,
   onAddUser,
   onRemoveUser,
   onApplySoul,
@@ -45,9 +55,9 @@ export function AgentForm({
   onOpenSkillInstallModal,
   onDelete,
 }: Props) {
-  const navigate = useNavigate();
   const { t } = useI18n();
   const { editingId, activeTab, isAdmin, form, currentUserId } = state;
+  const embedded = layout === "embedded";
 
   const canEdit = isAdmin || !editingId || (form.creator_id && form.creator_id === currentUserId);
 
@@ -55,75 +65,106 @@ export function AgentForm({
     (u: User) => !state.assignedUsers.some((a: User) => a.id === u.id),
   );
 
+  const hidden = new Set(hiddenTabs ?? []);
+  const shows = (tab: string) => !hidden.has(tab);
+  // A hidden tab must never stay selected — the trigger is gone, so the panel
+  // would render nothing at all.
+  const currentTab = shows(activeTab) ? activeTab : "config";
+
   return (
-    <div className="h-full min-h-0 min-w-0 flex flex-col bg-card">
-      <div className="border-b border-border px-6 py-4">
-        <span className="font-medium text-sm text-foreground">
-          {editingId ? t("agents.form.editAgent", { name: form.name }) : t("agents.form.newAgent")}
-        </span>
-      </div>
+    <div
+      className={
+        embedded
+          ? "flex min-w-0 flex-col overflow-hidden rounded-lg border border-border bg-card"
+          : "h-full min-h-0 min-w-0 flex flex-col bg-card"
+      }
+    >
+      {!embedded && (
+        <div className="border-b border-border px-6 py-4">
+          <span className="font-medium text-sm text-foreground">
+            {editingId
+              ? t("agents.form.editAgent", { name: form.name })
+              : t("agents.form.newAgent")}
+          </span>
+        </div>
+      )}
       <Tabs
-        value={activeTab}
-        onValueChange={(tab) => {
-          onSetState({ activeTab: tab as string });
-          if (editingId) {
-            void navigate({
-              to: "/settings/agents/$agentId/$tab",
-              params: { agentId: editingId, tab: tab as string },
-            });
-          }
-          if (tab === "users" && editingId) onLoadAssignedUsers(editingId);
-        }}
-        className="flex-1 flex flex-col min-h-0"
+        value={currentTab}
+        onValueChange={(tab) => onTabChange(tab as string)}
+        className={embedded ? "flex flex-col" : "flex-1 flex flex-col min-h-0"}
       >
         <TabsList
           variant="underline"
-          className="w-full justify-start px-6 border-b border-border rounded-none bg-transparent gap-2 h-11"
+          className={`w-full justify-start border-b border-border rounded-none bg-transparent gap-2 h-11 ${
+            embedded ? "px-4" : "px-6"
+          }`}
         >
-          <TabsTrigger value="config">{t("agents.tabs.config")}</TabsTrigger>
-          <TabsTrigger value="prompt">{t("agents.tabs.prompt")}</TabsTrigger>
-          <TabsTrigger value="skills">{t("agents.tabs.skills")}</TabsTrigger>
-          <TabsTrigger value="tools">{t("agents.tabs.tools")}</TabsTrigger>
-          <TabsTrigger value="advanced">{t("agents.tabs.advanced")}</TabsTrigger>
-          {isAdmin && <TabsTrigger value="users">{t("agents.tabs.users")}</TabsTrigger>}
+          {shows("config") && <TabsTrigger value="config">{t("agents.tabs.config")}</TabsTrigger>}
+          {shows("prompt") && <TabsTrigger value="prompt">{t("agents.tabs.prompt")}</TabsTrigger>}
+          {shows("skills") && <TabsTrigger value="skills">{t("agents.tabs.skills")}</TabsTrigger>}
+          {shows("tools") && <TabsTrigger value="tools">{t("agents.tabs.tools")}</TabsTrigger>}
+          {shows("advanced") && (
+            <TabsTrigger value="advanced">{t("agents.tabs.advanced")}</TabsTrigger>
+          )}
+          {isAdmin && shows("users") && (
+            <TabsTrigger value="users">{t("agents.tabs.users")}</TabsTrigger>
+          )}
         </TabsList>
-        <div className="flex-1 min-h-0 overflow-y-auto p-6 space-y-6">
-          <TabsContent value="config">
-            <ConfigTab state={state} onSetState={onSetState} />
-          </TabsContent>
-          <TabsContent value="prompt">
-            <PromptTab state={state} onSetState={onSetState} onApplySoul={onApplySoul} />
-          </TabsContent>
-          <TabsContent value="skills">
-            <SkillsTab
-              state={state}
-              onSetState={onSetState}
-              onSelectSkill={onSelectSkill}
-              onSaveSelectedSkill={onSaveSelectedSkill}
-              onDeleteSkill={onDeleteSkill}
-              onSelectSkillFile={onSelectSkillFile}
-              onDeleteSkillFile={onDeleteSkillFile}
-              onOpenSkillInstallModal={onOpenSkillInstallModal}
-            />
-          </TabsContent>
-          <TabsContent value="tools">
-            <ToolsTab state={state} />
-          </TabsContent>
-          <TabsContent value="advanced">
-            <AdvancedTab state={state} onSetState={onSetState} />
-          </TabsContent>
-          <TabsContent value="users">
-            <UsersTab
-              state={state}
-              availableUsers={availableUsers}
-              onSetState={onSetState}
-              onAddUser={onAddUser}
-              onRemoveUser={onRemoveUser}
-            />
-          </TabsContent>
+        <div
+          className={embedded ? "p-4 space-y-6" : "flex-1 min-h-0 overflow-y-auto p-6 space-y-6"}
+        >
+          {shows("config") && (
+            <TabsContent value="config">
+              <ConfigTab state={state} onSetState={onSetState} />
+            </TabsContent>
+          )}
+          {shows("prompt") && (
+            <TabsContent value="prompt">
+              <PromptTab state={state} onSetState={onSetState} onApplySoul={onApplySoul} />
+            </TabsContent>
+          )}
+          {shows("skills") && (
+            <TabsContent value="skills">
+              <SkillsTab
+                state={state}
+                onSetState={onSetState}
+                onSelectSkill={onSelectSkill}
+                onSaveSelectedSkill={onSaveSelectedSkill}
+                onDeleteSkill={onDeleteSkill}
+                onSelectSkillFile={onSelectSkillFile}
+                onDeleteSkillFile={onDeleteSkillFile}
+                onOpenSkillInstallModal={onOpenSkillInstallModal}
+              />
+            </TabsContent>
+          )}
+          {shows("tools") && (
+            <TabsContent value="tools">
+              <ToolsTab state={state} />
+            </TabsContent>
+          )}
+          {shows("advanced") && (
+            <TabsContent value="advanced">
+              <AdvancedTab state={state} onSetState={onSetState} />
+            </TabsContent>
+          )}
+          {shows("users") && (
+            <TabsContent value="users">
+              <UsersTab
+                state={state}
+                availableUsers={availableUsers}
+                onSetState={onSetState}
+                onAddUser={onAddUser}
+                onRemoveUser={onRemoveUser}
+              />
+            </TabsContent>
+          )}
         </div>
       </Tabs>
-      <div className="shrink-0 border-t border-border px-6 py-4 flex items-center justify-between gap-2">
+      <div
+        className={`shrink-0 border-t border-border py-4 flex items-center justify-between gap-2 ${
+          embedded ? "px-4" : "px-6"
+        }`}
+      >
         <div>
           {canEdit && editingId && onDelete && (
             <Button
@@ -137,9 +178,11 @@ export function AgentForm({
           )}
         </div>
         <div className="flex items-center gap-2">
-          <Button onClick={onCancel} variant="ghost" size="sm" className="cursor-pointer">
-            {t("common.cancel")}
-          </Button>
+          {onCancel && (
+            <Button onClick={onCancel} variant="ghost" size="sm" className="cursor-pointer">
+              {t("common.cancel")}
+            </Button>
+          )}
           {canEdit && (
             <Button onClick={onSave} size="sm" className="cursor-pointer">
               {editingId ? t("common.update") : t("common.create")}
