+++ web/src/features/agents/tabs/AdvancedTab.tsx
@@ -1,5 +1,5 @@
-import { normalizeSandbox } from "../AgentsPage";
-import type { AgentsPageState } from "../AgentsPage";
+import { normalizeSandbox } from "@/lib/queries/agent-settings";
+import type { AgentsPageState } from "../agent-detail-state";
 import { Textarea } from "@/components/ui/textarea";
 import { useI18n } from "@/lib/i18n";
