+++ web/src/features/agents/tabs/UsersTab.tsx
@@ -1,5 +1,5 @@
 import type { User } from "@/lib/types";
-import type { AgentsPageState } from "../AgentsPage";
+import type { AgentsPageState } from "../agent-detail-state";
 import { Button } from "@/components/ui/button";
 import { useI18n } from "@/lib/i18n";
