+++ web/src/lib/i18n/messages.ts
@@ -66,6 +66,17 @@ const en = {
   "profile.skills": "Skills",
   "profile.manageSkills": "Manage skills",
   "profile.noSkills": "No skills installed.",
+  "profile.about": "About",
+  "profile.aboutModel": "Model",
+  "profile.aboutScope": "Scope",
+  "profile.aboutScopeSystem": "All users",
+  "profile.aboutScopeRestricted": "Assigned users only",
+  "profile.aboutStatus": "Status",
+  "profile.aboutEnabled": "Enabled",
+  "profile.aboutDisabled": "Disabled",
+  "profile.aboutLastActive": "Last active",
+  "profile.configuration": "Configuration",
+  "profile.configurationDesc": "Model, prompt, tools and sandbox for this agent.",
   "profile.files": "Files",
   "profile.openFiles": "Open files",
   "profile.more": "More",
@@ -107,6 +118,8 @@ const en = {
   "common.copy": "Copy",
   "common.copied": "Copied",
   "common.update": "Update",
+  "common.saved": "Saved",
+  "common.deleted": "Deleted",
   "common.install": "Install",
   "common.uninstall": "Uninstall",
   "common.refresh": "Refresh",
@@ -337,6 +350,16 @@ const en = {
   "agents.skills.installed.success": "Skill installed",
   "agents.skills.uninstalled.success": "Skill uninstalled",
   "agents.skills.installFailed": "Failed to install skill",
+  "agents.detail.userAssigned": "User assigned",
+  "agents.detail.userRemoved": "User removed",
+  "agents.detail.discardChanges": "Discard unsaved changes?",
+  "agents.detail.confirmDeleteSkill": 'Delete skill "{{name}}"? This cannot be undone.',
+  "agents.detail.confirmDeleteFile": 'Delete file "{{name}}"?',
+  "agents.detail.skillRemoved": "Skill removed",
+  "agents.detail.chooseSkillFirst": "Choose a skill first",
+  "agents.detail.chooseZipFirst": "Choose a .zip file first",
+  "agents.detail.skillInstalled": "Installed: {{name}}",
+  "agents.detail.skillUploaded": "Uploaded: {{name}}",
   "agents.users.title": "Users",
   "agents.users.add": "Add User",
   "agents.users.noUsers": "No users assigned",
@@ -2091,6 +2114,17 @@ const zh: Record<MessageKey, string> = {
   "profile.skills": "技能",
   "profile.manageSkills": "管理技能",
   "profile.noSkills": "尚未安装技能。",
+  "profile.about": "关于",
+  "profile.aboutModel": "模型",
+  "profile.aboutScope": "范围",
+  "profile.aboutScopeSystem": "所有用户",
+  "profile.aboutScopeRestricted": "仅已分配用户",
+  "profile.aboutStatus": "状态",
+  "profile.aboutEnabled": "已启用",
+  "profile.aboutDisabled": "已停用",
+  "profile.aboutLastActive": "最近活跃",
+  "profile.configuration": "配置",
+  "profile.configurationDesc": "该智能体的模型、提示词、工具与沙箱。",
   "profile.files": "文件",
   "profile.openFiles": "打开文件",
   "profile.more": "更多",
@@ -2132,6 +2166,8 @@ const zh: Record<MessageKey, string> = {
   "common.copy": "复制",
   "common.copied": "已复制",
   "common.update": "更新",
+  "common.saved": "已保存",
+  "common.deleted": "已删除",
   "common.install": "安装",
   "common.uninstall": "卸载",
   "common.refresh": "刷新",
@@ -2352,6 +2388,16 @@ const zh: Record<MessageKey, string> = {
   "agents.skills.installed.success": "技能已安装",
   "agents.skills.uninstalled.success": "技能已卸载",
   "agents.skills.installFailed": "安装技能失败",
+  "agents.detail.userAssigned": "已分配用户",
+  "agents.detail.userRemoved": "已移除用户",
+  "agents.detail.discardChanges": "放弃未保存的修改？",
+  "agents.detail.confirmDeleteSkill": "删除技能“{{name}}”？此操作无法撤销。",
+  "agents.detail.confirmDeleteFile": "删除文件“{{name}}”？",
+  "agents.detail.skillRemoved": "已移除技能",
+  "agents.detail.chooseSkillFirst": "请先选择一个技能",
+  "agents.detail.chooseZipFirst": "请先选择 .zip 文件",
+  "agents.detail.skillInstalled": "已安装：{{name}}",
+  "agents.detail.skillUploaded": "已上传：{{name}}",
   "agents.users.title": "用户",
   "agents.users.add": "添加用户",
   "agents.users.noUsers": "未分配用户",
