+++ web/src/lib/route-search.ts
@@ -1,6 +1,12 @@
 export interface MemorySearch {
   knowledge?: "removed";
+  /** Selected tab of the agent profile's configuration editor. */
+  ctab?: AgentConfigTab;
 }
+
+export type AgentConfigTab = "config" | "prompt" | "tools" | "advanced" | "users";
+
+const AGENT_CONFIG_TABS = new Set<string>(["config", "prompt", "tools", "advanced", "users"]);
 
 export interface SkillsSearch {
   new?: boolean;
@@ -13,7 +19,12 @@ const SKILL_SOURCES = new Set(["installed", "market", "manual"]);
 const SKILL_SCOPES = new Set(["project", "user", "agent", "system"]);
 
 export function validateMemorySearch(search: Record<string, unknown>): MemorySearch {
-  return search.knowledge === "removed" ? { knowledge: "removed" } : {};
+  return {
+    ...(search.knowledge === "removed" ? { knowledge: "removed" as const } : {}),
+    ...(AGENT_CONFIG_TABS.has(search.ctab as string)
+      ? { ctab: search.ctab as AgentConfigTab }
+      : {}),
+  };
 }
 
 export function validateSkillsSearch(search: Record<string, unknown>): SkillsSearch {
