+++ web/src/routes/_app/settings/agents.$agentId.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { loadAgentsSettingsData } from "@/features/agents/AgentsPage";
+import { loadAgentsSettingsData } from "@/lib/queries/agent-settings";
 
 export const Route = createFileRoute("/_app/settings/agents/$agentId")({
   loader: ({ params: { agentId } }) => loadAgentsSettingsData(agentId),
